Create a developer user
It is preferable that we don't constantly use the root user while inside
containers created from this image. This is the first commit on the path
to achieving this.

I have increased the major version since it will require new
instructions to start the containers with permissions mapped to the host
operating system. Without following these new instructions the user will
encounter permissions issues.

I have not yet figured out how to do this mapping though so this version
is not ready to be tagged just yet.