Install neovim using apt rather than appimage
This is not the only change here. I've also added another call to
`RUN +UpdateRemotePlugins +qall` in the Dockerfile and there is a new
appendage on the vim.init file (in the dotfiles submodule) which is
running the `./install` of nvm-typescript plugin properly. All the info
is in cluded in this an similar forum posts: https://github.com/mhartington/nvim-typescript/issues/139

So the fix may have been those changes or the installing of neovim
through apt. Either way, I'm happy with an apt install so I'm keeping
it.