Tear out the lesser required things from this file
It's a shame to not be able to run docker from inside the container, but
if there is one thing we know about the environment which this container
is being run in, it's that it has docker installed. It's reasonable to
carry out the docker interactions through that.

I've also removed all the generators which I don't use often enough to
tollerate sitting through their installation each time I'm making
changes.