Add simple .bashrc and and CMD entry of bash shell
This .bashrc is a simplified version of the type which is deployed with
a standard Ubuntu environment. I really only want it in order to load
the .bash_aliases file and I've left in some of the prompt decorations
since they're there.

I have not added the .bashrc to the dotfiles project because there is no
general purpose use for it. It's not a portably dotfile which I might
want to use elsewhere. At least, that's the way I see it right now.