Download neovim 0.5 in place of 0.4
Since this is only available as an appimage package at the moment I have
had to move back to a Ubuntu base image from Alpine. There might be a
way to get it working on Alpine, but I'm not interested in solving that
problem yet. I just want to see what neovim 0.5 is like in a container
first.