Add tsserver for neovim LSP
I'm not sure about the single line install of nvm npm and typescript.
Its the only way I've found to get the updates of one bash shell command
to be carried through in to the other. If they are each individual RUN
commands then `nvm` isn't available in the next shell even if the
`~/.bashrc` is sourced.

I've tried running `SHELL ["/bin/bash", "-c"]` to change the default
shell and it still appears that shell contexts are not shared between
commands. I've got a lot to learn about Dockerfiles still.