Install bat for Telescope preview (working?)
The `bat` command is working now, but unfortunately I can't see any
evidence that it's having an effect in the previewer. It's unclear what
it's supposed to be doing so I might have to return to this. It seems
like treesitter is equally as important for syntax highlighting so
perhaps I'm expecting something to happen that I shouldn't be. The
Telescope docs aren't as clear as they could be yet. It's still an early
release.