Completion plugin from dotfiles and improve install
This approach ensures that all treesitter parsers will be installed in
to the image before the first container is created, regardless of how
many parsers are configured in the actual vimrc. This project sits
independently from the dotfiles submodule in many ways, and the
preinstallation of all of the parsers is one guarantee I would like the
image to provide.

It is important too since we don't want the nvim instance in the
container to spend the first few seconds of every iteration of it's life
doing the same things and momentarily leaving files unparsed.