Add instruction for nerd fonts in windows for work
Maybe many more work too. I tried the Droid Sans Mono... one from the curl request above, but it didn't work. Perhaps the difference is that this one is ttf.

The method I used to import was to just go to the "Fonts" menu in the "Control panel" and drag the ttf file in. It was immediately available in the properties of WSL after that.