Simplify Dockerfile definition
This may be less feature rich than it was before, and there might be
some situations where it doesn't build or run well now that I've removed
the proxy settings.

However, it does what I need it to right now, it has a much better
version of neovim in it, and it can be attached as a dev container to
other docker volumes now.
imagining setting