Remove unused resources
Arguably there is a place for some of the functionality removed here.
And arguably the neovim conig should be part of the dotfiles project.

However, the dotfiles project is showing it's age and requires a lot of
cleanup itself. I should get to that but I'm still proving the concept
of the new dev-env setup.

Also, I'll only be reintroducing other functionality as the need arises,
so some of the things removed here might find their way back, but not
without a reason.