Add a simple folder to edit the project from
This means that the development environment can be edited from inside
itself. The tools to do the editing are already available so there's no
need for volume mounting in this usecase.

The README.md update in this commit is not satisfactory, but it's enough
for me to organise my thoughts for now. The README file needs to be
reorganised in consideration of this change.