Remove unhelpful labels
The purpose of semver in this image was removed a while ago really. Now
that we're not getting the run command from the image lables the semver
is just something to forgot to keep up to date. It's also redundant
since the tag id is part of the image metadata.