Handle lists of 1 better
Previously when only a single item was returned, the variable would be
assigned as a string rather than an array. Subsequently, when index
extraction is used to display the list of items then it would only show
the first letter. This forces the result set to be an array leading to
consistent behaviour.