Switch to plaintext GCM storage method
This isn't perfect, but for the contexts which I'll be using the GCM it
is acceptable. If someone gains access to my laptop and files then I'm
in big trouble already.

Plus, the intended workflow is that I will give myself an access token
per project, so compromising an access code from one of the neovim
containers will only reveal an access token for a single project with
Developer permissions.